refactor the rest of the jobController functions
