+++ src/controllers/jobs/index.ts
@@ -4,3 +4,6 @@ export * from './getJobs.js'
 export * from './jobUtils.js'
 export * from './autoRg.js'
 export * from './scoperStatus.js'
+export * from './updateJob.js'
+export * from './downloadJobResults.js'
+export * from './getLogForStep.js'
